requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.20.0
click==8.1.5
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
lazy-model==0.0.5
Mako==1.2.4
MarkupSafe==2.1.3
motor==3.2.0
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.11
pydantic-core==2.3.0
PyJWT==2.8.0
pymongo==4.4.1
python-decouple==3.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
toml==0.10.2
typing-extensions==4.7.1
uvicorn==0.23.0
